19 February 2009 merge from tamarin-redux to tamarin-central
There were 295 entries in the log between the November 2008 merge and this one.
Highlights:

  SPARC back-end for NanoJIT
  x86_64 back-end for NanoJIT
  PPC64 back-end for NanoJIT
  String class ported from tamarin-tracing
  Verifier work: documentation; rework the verifier as a pipeline in
    order to separate translation/jitting from verification
  Added a prototype abc assembler
  Significant work on portability, separating the VM code from the
    platform, using standard C++, etc
  Significant mmgc work on cleanup, profiling, and out-of-memory handling
  Code size reports
  Many bug fixes, optimizations, and cleanup throughout

+++ extensions/ST_avmplus_peephole.cpp
@@ -97,8 +97,8 @@ void ST_avmplus_peephole::test0() {
      t->emitOp1(WOP_getlocal, 65536);
      t->emitOp1(WOP_getlocal, 7);
      t->emitOp1(WOP_getlocal, 6);
-     uint32_t* code;
-     uint32_t len = t->epilogue(&code);
+     uintptr_t* code;
+uint32_t len = (uint32_t)t->epilogue(&code);
 
 verifyPass(len == 6, "len == 6", __FILE__, __LINE__);
 verifyPass(code[0] == NEW_OPCODE(WOP_get2locals), "code[0] == NEW_OPCODE(WOP_get2locals)", __FILE__, __LINE__);
